Validations added in the Reader class and error handling added when the reader cannot be added.

+++ src/main/java/com/example/libray_rest_api/libray_rest_api/controller/ReaderController.java
@@ -3,6 +3,7 @@ package com.example.libray_rest_api.libray_rest_api.controller;
 import com.example.libray_rest_api.libray_rest_api.domain.Dto.ReaderDto;
 import com.example.libray_rest_api.libray_rest_api.domain.Reader;
 import com.example.libray_rest_api.libray_rest_api.domain.exception.ReaderNotFound;
+import com.example.libray_rest_api.libray_rest_api.domain.exception.ReaderNotPossibleToAdd;
 import com.example.libray_rest_api.libray_rest_api.mapper.ReaderMapper;
 import com.example.libray_rest_api.libray_rest_api.service.ReaderDbService;
 import jakarta.transaction.Transactional;
@@ -30,25 +31,27 @@ public class ReaderController {
 
     @PostMapping(value = "/add", consumes = MediaType.APPLICATION_JSON_VALUE)
     public void createReader(@RequestBody ReaderDto readerDto) {
-        Reader reader = readerMapper.mapToReader(new ReaderDto(
-                null,
-                readerDto.getFirstname(),
-                readerDto.getLastname(),
-                LocalDate.now()
-        ));
-        readerDbService.saveToDataBase(reader);
+        try {
+            Reader reader = readerMapper.mapToReader(new ReaderDto(
+                    null,
+                    readerDto.getFirstname(),
+                    readerDto.getLastname(),
+                    LocalDate.now()
+            ));
+            readerDbService.saveToDataBase(reader);
+        } catch (Exception e) {
+            throw new ReaderNotPossibleToAdd();
+        }
     }
 
     @Transactional
     @DeleteMapping(value = "/deleteById/{reader_id}")
-    public ResponseEntity<Void> deleteReaderById(@PathVariable Long reader_id) throws ReaderNotFound {
-        try {
+    public ResponseEntity<Void> deleteReaderById(@PathVariable Long reader_id) {
+        if (readerDbService.existsByIdFromDataBase(reader_id)) {
             readerDbService.deleteByIdFromDataBase(reader_id);
-        } catch (ReaderNotFound readerNotFound) {
+        } else {
             throw new ReaderNotFound();
         }
         return ResponseEntity.ok().build();
     }
-
-
 }
